+++ src/main/java/nextstep/favorite/application/FavoriteService.java
@@ -10,8 +10,7 @@ import nextstep.line.LineRepository;
 import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
-import nextstep.path.Path;
-import nextstep.path.PathFinder;
+import nextstep.path.SubwayMap;
 import nextstep.station.Station;
 import nextstep.station.StationRepository;
 import org.springframework.stereotype.Service;
@@ -40,7 +39,7 @@ public class FavoriteService {
         Station targetStation = stationRepository.findById(request.getTarget()).orElseThrow(() -> new SubwayException("존재하지 않는 역입니다."));
         List<Line> lines = lineRepository.findAll();
 
-        PathFinder pathFinder = new PathFinder(lines);
+        SubwayMap pathFinder = new SubwayMap(lines);
         pathFinder.isValidateRoute(sourceStation, targetStation);
 
         Favorite favorite = new Favorite(sourceStation, targetStation, member);
