+++ src/test/java/nextstep/path/PathFinderTest.java
@@ -59,8 +59,8 @@ public class PathFinderTest {
     @DisplayName("최단 거리 경로를 조회한다")
     @Test
     void findPathByDistance() {
-        PathFinder pathFinder = new PathFinder(Arrays.asList(이호선, 신분당선, 삼호선));
-        Path path = pathFinder.findPath(교대역, 양재역, PathType.DISTANCE);
+        SubwayMap subwayMap = new SubwayMap(Arrays.asList(이호선, 신분당선, 삼호선));
+        Path path = subwayMap.findPath(교대역, 양재역, PathType.DISTANCE);
 
         assertThat(path.getStations()).containsExactly(교대역, 남부터미널역, 양재역);
         assertThat(path.getDistance()).isEqualTo(5);
@@ -70,8 +70,8 @@ public class PathFinderTest {
     @DisplayName("최단 시간 경로를 조회한다")
     @Test
     void findPathByDuration() {
-        PathFinder pathFinder = new PathFinder(Arrays.asList(이호선, 신분당선, 삼호선));
-        Path path = pathFinder.findPath(교대역, 양재역, PathType.DURATION);
+        SubwayMap subwayMap = new SubwayMap(Arrays.asList(이호선, 신분당선, 삼호선));
+        Path path = subwayMap.findPath(교대역, 양재역, PathType.DURATION);
 
         assertThat(path.getStations()).containsExactly(교대역, 강남역, 양재역);
         assertThat(path.getDistance()).isEqualTo(20);
@@ -81,9 +81,9 @@ public class PathFinderTest {
     @DisplayName("출발역과 도착역이 같은 경우 에러를 반환한다.")
     @Test
     void validateEqualsStation() {
-        PathFinder pathFinder = new PathFinder(Arrays.asList(이호선, 신분당선, 삼호선));
+        SubwayMap subwayMap = new SubwayMap(Arrays.asList(이호선, 신분당선, 삼호선));
 
-        assertThatThrownBy(() -> pathFinder.findPath(교대역, 교대역, PathType.DISTANCE))
+        assertThatThrownBy(() -> subwayMap.findPath(교대역, 교대역, PathType.DISTANCE))
                 .isInstanceOf(SubwayException.class)
                 .hasMessageContaining("출발역과 도착역이 같습니다.");
     }
@@ -91,9 +91,9 @@ public class PathFinderTest {
     @DisplayName("출발역과 도착역이 연결이 되어 있지 않은 경우 에러를 반환한다.")
     @Test
     void validatePathExists() {
-        PathFinder pathFinder = new PathFinder(Arrays.asList(이호선, 수인분당선));
+        SubwayMap subwayMap = new SubwayMap(Arrays.asList(이호선, 수인분당선));
 
-        assertThatThrownBy(() -> pathFinder.findPath(강남역, 수내역, PathType.DISTANCE))
+        assertThatThrownBy(() -> subwayMap.findPath(강남역, 수내역, PathType.DISTANCE))
                 .isInstanceOf(SubwayException.class)
                 .hasMessageContaining("출발역과 도착역이 연결이 되어 있지 않습니다.");
     }
@@ -101,9 +101,9 @@ public class PathFinderTest {
     @DisplayName("존재하지 않은 출발역이나 도착역을 조회 할 경우 에러를 반환한다.")
     @Test
     void validateStationExists() {
-        PathFinder pathFinder = new PathFinder(Arrays.asList(이호선));
+        SubwayMap subwayMap = new SubwayMap(Arrays.asList(이호선));
 
-        assertThatThrownBy(() -> pathFinder.findPath(강남역, 수내역, PathType.DISTANCE))
+        assertThatThrownBy(() -> subwayMap.findPath(강남역, 수내역, PathType.DISTANCE))
                 .isInstanceOf(SubwayException.class)
                 .hasMessageContaining("존재하지 않은 역입니다.");
     }
